refactor: 이미지 변환 함수 utils 폴더로 분리
