+++ src/utils/EventBus.ts
@@ -1,20 +1,22 @@
-const EventBus = {
-  listeners: {} as Record<string, Array<(data: any) => Promise<void> | void>>,
+type EventCallback = (data: string | null) => Promise<void> | void;
 
-  subscribe: (event: string, callback: (data: any) => Promise<void> | void) => {
+const EventBus = {
+  listeners: {} as Record<string, EventCallback[]>,
+
+  subscribe: (event: string, callback: EventCallback) => {
     if (!EventBus.listeners[event]) {
       EventBus.listeners[event] = [];
     }
     EventBus.listeners[event].push(callback);
   },
 
-  unsubscribe: (event: string, callback: (data: any) => Promise<void> | void) => {
+  unsubscribe: (event: string, callback: EventCallback) => {
     if (EventBus.listeners[event]) {
       EventBus.listeners[event] = EventBus.listeners[event].filter((listener) => listener !== callback);
     }
   },
 
-  publish: (event: string, data: any) => {
+  publish: (event: string, data: string | null) => {
     if (EventBus.listeners[event]) {
       const promises = EventBus.listeners[event].map((callback) => callback(data));
       return Promise.all(promises);
